core: enhanced noopener output (#5857)
